Expose the type of signal numbers as SigNo
Before this change, users of signal-hook might have had to directly
reference libc::c_int in order to be agnostic to the size of the
integers used to represent signal numbers.  This is unfortunate.

With the advent of nix, it's reasonable for programs to stay away
from any direct use of libc -- and it'd be nice if using signal-hook
didn't require the reintroduction of a direct dependency on libc.