Add support for smol async runtime
The signal-hook trait didn't support the smol runtime yet. Furthermore
async_std uses smol since 1.6.0 so this changes adds indirect support for
this runtime, too.

This changes focuses onto the interator part of signal-hooks. It adds a
generic adapter which implements the Stream trait of the futures library.
The adapter takes an AsyncRead object which is used to schedule the read
end of the pipe with the runtime. It also adds an into_stream method
to Signals which creates a smoll Async instance for the read end.