Add some tests for the Signals struct and fix a bug that showed up
The wait method of the Signal struct assumed there was no way it could
recieve None from the call to poll_pending method. But this wasn't true
in the case the instance was already closed. In this case the
poll_pending method returns None.

Now we return self.pending() in this case because this is the same
behaviour as before the refactorings for the async crate split.