Move adding of signals and closing of the iterator into a shareable handle (#49)
* Move adding of signals and closing of the iterator into a handle

The handle can be cloned and moved between threads. Due to this the
Signal struct doesn't have to implement the Clone trait to allow adding
signals and closing from other treads. This allows to move the reading
end of the pipe to the SignalDelivery struct and get mutable references
to it.

Using inherited mutability instead of interior mutability then allows to
guarantee that there is only one thread reading from the pipe at the
same time. This simplifies the close mechanic and some methods.