Don't leak a thread in tests
Don't leave a thread waiting for signals behind. First, it's a bad form.
Second, it actually did receive the signal during some more tests and
the thread panics in the background, making noise on the console. And
the panic was not propagated anywhere.

So tell it to terminate, wait for it and make sure any legitimate panics
are propagated to the main thread.