Bump serial_test dev-dependency from v0.5 to v3
Version 0.5 is five years old at this point. It looks like there
were no API changes between v0.5 and v3 that affect this project,
so the dependency can just be bumped.

This bump also results in a lot of obsolete versions of crates
being dropped from the project's dependency tree.