Replace internal maps with unsorted Vecs
HashMap and BTreeMap are overkill in this context. Unsorted vectors are
plenty fast enough and the necessary collection interfaces are
straightforward to implement. This change has two benefits.

First, it improves binary size. For the `print` example from signal-hook
in release mode, the .text section shrinks by about 18 KiB and overall
file size shrinks by about 30 KiB. That's roughly a 6% reduction in both
metrics.

Second, the simpler data structures make it more obvious that the signal
handler only does async-signal-safe operations. In particular, the
default HashMap has a `RandomState`, which can access TLS, do dlsym
lookups, open and read from files, etc. depending on the platform. I
don't think that's a problem for the hash table *lookup* done in the
signal handler since that shouldn't touch the `RandomState`, but it's a
bit subtle and the standard library doesn't make any guarantees about
this. Avoiding hash maps entirely removes the need to think about it.

Performance notes:

* (Un-)registering actions does an insert/remove by ActionId, which is
  asymptotically slower with this PR. However, (un-)registering is a
  slow operation and should be done rarely. Besides locking, it always
  clones the entire `SignalData`, so it already takes O(n) time when
  there's n actions registered across all symbols.
* The signal handler looks up the `Slot` by signal number, which is
  asymptotically slower with this PR. However, there's only a very small
  constant number of signals, so asymptotics don't matter.
* After looking up the right `Slot`, the signal handler only iterates
  sequentially over the actions, it doesn't do any lookups by ActionId.
* For a simple microbenchmark that registers one action each for 20
  signals and then raises one signal 100k times, this implementation
  appears to be slightly *faster* regardless of which signal is raised.